refactor(std::layout): split lib/stdlib/layout.ts into per-concern files
The implementation file was a 1030-line grab-bag mixing ANSI / SGR
primitives, the Block algebra, border framing, the node tree, leaf
renderers, axis composition, box, and the table renderer. Split it
into one file per concern under `lib/stdlib/layout/`:

  ansi.ts    — Style, SGR / color, visualWidth, stripAnsi, styledWrapper
  block.ts   — Block, Align, pad, padLine, beside, above, styled
  border.ts  — BORDER_CHARS, BorderStyle, BorderChars, resolveBorderStyle,
                bordered, BorderOpts, buildTopEdge, minWidthForTitle
  nodes.ts   — NodeType, LayoutNode, Cell, ColumnSpec, styleOf,
                alignedTextBlock, LEAF_RENDERERS (text/raw/space/hline/vline)
  axis.ts    — composeRow, composeColumn, composeAxis, AXIS_OPS
  box.ts     — composeBox
  table.ts   — _coerceCell, _validateTable, ColumnLayout, _applyColumnFg,
                _computeColumnLayouts, _composeRowBlock, _composeDividerLine,
                _styleHeaderCell, _composeCaption, composeTable + helpers
  render.ts  — RENDERERS dispatch table + renderNode + render + _render

Dependency graph is a clean DAG with one benign cycle: render.ts
imports each container compose function; the container files import
`renderNode` from render.ts (used inside function bodies only, so
module loading completes before any call). No other cycles.

`lib/stdlib/layout.ts` is now a 40-line public re-exporter. The
package subpath import `agency-lang/stdlib-lib/layout.js` still
resolves to `dist/lib/stdlib/layout.js` so the Agency side and the
test helper (`tests/helpers/tryRender.js`) keep working unchanged.
The test file's `./layout.js` import + `_internal` surface are
preserved verbatim.

No behavior change. 143 unit tests + 25 integration tests + the
structural linter all stay green.

Amp-Thread-ID: https://ampcode.com/threads/T-019e95e9-4faf-70c9-893c-d31ea0f6e8b8
