+++ packages/agency-lang/lib/stdlib/layout/table.ts
@@ -1,0 +1,401 @@
+// std::layout — table renderer.
+//
+// Pipeline:
+//   1. _validateTable          — structural check + cell coercion
+//   2. _computeColumnLayouts   — measure each cell, build per-column
+//                                 layout (width / align / padding / fg)
+//   3. composeTable            — declaratively assemble a section list
+//                                 (row blocks interleaved with dividers),
+//                                 wrap the lot in the outer frame + caption
+//
+// `ColumnLayout` is the single source of truth for column geometry
+// and style. Anything that needs to know "how wide is column N
+// including padding?" or "what's the inner width of the cell grid?"
+// derives from it via `_innerTableWidth`. The cell-renderer
+// (`_composeRowBlock`) and the divider-renderer (`_composeDividerLine`)
+// both consume the same `ColumnLayout[]`, so a change to one stays in
+// sync with the other automatically.
+
+import { Style, styledWrapper, visualWidth } from "./ansi.js";
+import { Align, Block, above, beside, pad, padLine, styled } from "./block.js";
+import {
+  BORDER_CHARS, BorderChars, BorderStyle, buildTopEdge, minWidthForTitle,
+  resolveBorderStyle,
+} from "./border.js";
+import { Cell, ColumnSpec, LayoutNode } from "./nodes.js";
+import { renderNode } from "./render.js";
+
+export { Cell, ColumnSpec };
+
+export function _coerceCell(cell: unknown): LayoutNode {
+  if (typeof cell === "string") {
+    return { type: "text", attrs: { content: cell }, children: [] };
+  }
+  if (
+    cell !== null &&
+    typeof cell === "object" &&
+    Object.hasOwn(cell, "type") &&
+    typeof (cell as { type: unknown }).type === "string" &&
+    Array.isArray((cell as { children: unknown }).children)
+  ) {
+    return cell as LayoutNode;
+  }
+  throw new Error(
+    `std::layout.table: cell must be string or LayoutNode, got ${
+      cell === null ? "null" : typeof cell
+    }`,
+  );
+}
+
+export type ValidatedTable = {
+  // `header` is `LayoutNode[]` (empty when absent) — not `[] | null` —
+  // for symmetry with `body` and `footer`. Callers iterate uniformly
+  // without a null guard.
+  header: LayoutNode[];
+  body:   LayoutNode[][];
+  footer: LayoutNode[][];
+  columnCount: number;
+};
+
+// Structural sanity for a `table` node. Throws on:
+//   * all three sections empty/unset
+//   * any present section whose row length disagrees with the column
+//     count (derived from `columns` if set, else `header`, else first
+//     body row, else first footer row)
+//
+// On success returns the three sections with every cell coerced to a
+// `LayoutNode`, plus the resolved column count.
+export function _validateTable(attrs: Record<string, unknown>): ValidatedTable {
+  const rawHeader = attrs.header;
+  const rawBody   = attrs.body   ?? [];
+  const rawFooter = attrs.footer ?? [];
+
+  if (rawHeader != null && !Array.isArray(rawHeader)) {
+    throw new Error(
+      `std::layout.table: header must be an array of cells, got ${typeof rawHeader}`,
+    );
+  }
+  const checkSection = (val: unknown, name: string): unknown[][] => {
+    if (!Array.isArray(val)) {
+      throw new Error(
+        `std::layout.table: ${name} must be an array of rows, got ${typeof val}`,
+      );
+    }
+    val.forEach((row, i) => {
+      if (!Array.isArray(row)) {
+        throw new Error(
+          `std::layout.table: ${name} row ${i} must be an array of cells, got ${typeof row}`,
+        );
+      }
+    });
+    return val as unknown[][];
+  };
+  const header = rawHeader as unknown[] | null | undefined;
+  const body   = checkSection(rawBody,   "body");
+  const footer = checkSection(rawFooter, "footer");
+
+  if (!header && body.length === 0 && footer.length === 0) {
+    throw new Error(
+      "std::layout.table: at least one of header / body / footer must be set",
+    );
+  }
+
+  const columns = attrs.columns as ColumnSpec[] | null | undefined;
+  let columnCount: number;
+  if (columns && columns.length > 0) columnCount = columns.length;
+  else if (header)                   columnCount = header.length;
+  else if (body.length > 0)          columnCount = body[0].length;
+  else                                columnCount = footer[0].length;
+
+  const checkRow = (row: unknown[], label: string) => {
+    if (row.length !== columnCount) {
+      throw new Error(
+        `std::layout.table: ${label} has ${row.length} cells, expected ${columnCount}`,
+      );
+    }
+  };
+  if (header) checkRow(header, "header");
+  body.forEach(  (r, i) => checkRow(r, `body row ${i}`));
+  footer.forEach((r, i) => checkRow(r, `footer row ${i}`));
+
+  return {
+    header: header ? header.map(_coerceCell) : [],
+    body:   body.map(row   => row.map(_coerceCell)),
+    footer: footer.map(row => row.map(_coerceCell)),
+    columnCount,
+  };
+}
+
+type ColumnLayout = {
+  width:       number;   // measured content width (after minWidth)
+  align:       Align;
+  cellPadding: number;
+  fgColor:     string;   // "" = none
+};
+
+// Inject the column's `fgColor` as the cell's `fgColor` only when the
+// cell is a `text` leaf without its own color. Cell-level fg always
+// wins; non-text leaves (raw / row / nested containers) are passed
+// through untouched.
+function _applyColumnFg(c: LayoutNode, colFg: string | undefined): LayoutNode {
+  if (!colFg || c.type !== "text") return c;
+  if (typeof c.attrs.fgColor === "string" && c.attrs.fgColor !== "") return c;
+  return { ...c, attrs: { ...c.attrs, fgColor: colFg } };
+}
+
+function _computeColumnLayouts(
+  rows: LayoutNode[][],
+  columnCount: number,
+  columns: ColumnSpec[],
+  cellPadding: number,
+): { layouts: ColumnLayout[]; cellBlocks: Block[][] } {
+  const styledRows = rows.map((row) =>
+    row.map((cell, c) => _applyColumnFg(cell, columns[c]?.fgColor)),
+  );
+  const cellBlocks = styledRows.map((row) => row.map(renderNode));
+  const layouts: ColumnLayout[] = [];
+  for (let c = 0; c < columnCount; c++) {
+    const measured = cellBlocks.reduce(
+      (m, row) => Math.max(m, row[c].width), 0,
+    );
+    const spec = columns[c] ?? {};
+    layouts.push({
+      width:       Math.max(measured, spec.minWidth ?? 0),
+      align:       spec.align ?? "start",
+      cellPadding,
+      fgColor:     spec.fgColor ?? "",
+    });
+  }
+  return { layouts, cellBlocks };
+}
+
+function _innerTableWidth(layouts: ColumnLayout[], columnDividers: boolean): number {
+  const cellsW = layouts.reduce((s, l) => s + l.width + 2 * l.cellPadding, 0);
+  return cellsW + (columnDividers ? Math.max(0, layouts.length - 1) : 0);
+}
+
+// Pad a cell to its column's content width with the column's
+// horizontal alignment, then add `cellPadding` spaces on each side.
+// The two-step pad keeps alignment relative to the *content* width
+// (not the padded width), so `align: "end"` still right-edges at the
+// column boundary, not at the outer cellPadding edge.
+function _layoutCell(block: Block, layout: ColumnLayout, rowHeight: number): Block {
+  const aligned = pad(block, layout.width, rowHeight, layout.align, "start");
+  if (layout.cellPadding <= 0) return aligned;
+  const padStr = " ".repeat(layout.cellPadding);
+  return Block.of(aligned.lines.map((l) => padStr + l + padStr));
+}
+
+// Build a single row block (cells joined, no outer side borders).
+function _composeRowBlock(
+  cells: Block[],
+  layouts: ColumnLayout[],
+  columnDividers: boolean,
+  dividerChar: string,
+  wrapBorder: (s: string) => string,
+): Block {
+  const rowHeight   = cells.reduce((m, b) => Math.max(m, b.height), 1);
+  const paddedCells = cells.map((b, c) => _layoutCell(b, layouts[c], rowHeight));
+  if (!columnDividers || paddedCells.length <= 1) {
+    return paddedCells.reduce(beside, Block.empty());
+  }
+  const dividerBlock = Block.of(Array(rowHeight).fill(wrapBorder(dividerChar)));
+  return paddedCells.reduce<Block>(
+    (acc, cell, i) =>
+      i === 0 ? cell : beside(beside(acc, dividerBlock), cell),
+    Block.empty(),
+  );
+}
+
+// Build a horizontal section divider with the proper junction
+// characters: `├` and `┤` on the side borders, `┼` at every column
+// crossing. When `innerWidth` exceeds the natural cell grid (e.g.
+// because a wide title widened the table), the extra width is filled
+// with plain `h` chars on the right.
+function _composeDividerLine(
+  layouts: ColumnLayout[],
+  columnDividers: boolean,
+  innerWidth: number,
+  ch: BorderChars,
+  wrapBorder: (s: string) => string,
+): Block {
+  const runs    = layouts.map((l) => ch.h.repeat(l.width + 2 * l.cellPadding));
+  const inner   = columnDividers ? runs.join(ch.cross) : runs.join("");
+  const padding = ch.h.repeat(Math.max(0, innerWidth - visualWidth(inner)));
+  return Block.of(wrapBorder(ch.leftTee + inner + padding + ch.rightTee));
+}
+
+// Tags that act as explicit opt-outs of header auto-bold. Agency's
+// `text()` constructor always serializes `bold: false` by default, so
+// treating `bold === false` as "set" would mean no `text()` cell ever
+// got the auto-bold — only bare strings would. We only treat
+// `bold === true` as a "do not touch" signal; any other explicit
+// modifier on the leaf (italic / dim / underline / fgColor / bgColor)
+// also opts out — the caller's styling wins.
+function _hasExplicitTextStyle(attrs: Record<string, unknown>): boolean {
+  return attrs.bold === true
+      || attrs.italic === true
+      || attrs.dim === true
+      || attrs.underline === true
+      || (typeof attrs.fgColor === "string" && attrs.fgColor !== "")
+      || (typeof attrs.bgColor === "string" && attrs.bgColor !== "");
+}
+
+function _styleHeaderCell(c: LayoutNode): LayoutNode {
+  // Only auto-bold `text` cells with no explicit modifier.
+  // Pre-built nodes (`raw`, `row`) and `text` with any explicit style
+  // modifier are passed through verbatim — the caller's styling wins.
+  if (c.type !== "text") return c;
+  if (_hasExplicitTextStyle(c.attrs)) return c;
+  return { ...c, attrs: { ...c.attrs, bold: true } };
+}
+
+// Wrap a row block's lines with the outer `│` side borders, padding
+// to `innerWidth` so the right edge lines up with the rest of the
+// table.
+function _wrapRowSides(
+  rowBlock: Block,
+  innerWidth: number,
+  ch: BorderChars,
+  wrapBorder: (s: string) => string,
+): Block {
+  const side = wrapBorder(ch.v);
+  return Block.of(
+    rowBlock.lines.map((l) => side + padLine(l, innerWidth, "start") + side),
+  );
+}
+
+// Declarative interleave of `row` and `divider` markers for one
+// section. Returns an empty array if `rows` is empty (so the caller
+// can simply concat the three sections without separator guards).
+type SectionPart = { kind: "row"; cells: Block[] } | { kind: "divider" };
+
+function _interleaveRows(
+  rows: Block[][],
+  rowDividers: boolean,
+): SectionPart[] {
+  return rows.flatMap((cells, i): SectionPart[] => {
+    const row: SectionPart = { kind: "row", cells };
+    if (rowDividers && i < rows.length - 1) {
+      return [row, { kind: "divider" }];
+    }
+    return [row];
+  });
+}
+
+// Glue header / body / footer into a single section list, inserting
+// header- and footer-dividers only where there is something on both
+// sides to separate.
+function _buildSectionParts(
+  header: Block[][],
+  body:   Block[][],
+  footer: Block[][],
+  opts: { headerDivider: boolean; footerDivider: boolean; rowDividers: boolean },
+): SectionPart[] {
+  const sections: SectionPart[][] = [
+    _interleaveRows(header, opts.rowDividers),
+    _interleaveRows(body,   opts.rowDividers),
+    _interleaveRows(footer, opts.rowDividers),
+  ];
+  const useHeaderDivider = opts.headerDivider && header.length > 0 && (body.length + footer.length) > 0;
+  const useFooterDivider = opts.footerDivider && footer.length > 0 && body.length > 0;
+
+  const separators: (SectionPart[] | null)[] = [
+    null,
+    useHeaderDivider ? [{ kind: "divider" }] : null,
+    useFooterDivider ? [{ kind: "divider" }] : null,
+  ];
+  return sections.flatMap((parts, i) => [...(separators[i] ?? []), ...parts]);
+}
+
+// Centered, dim caption row. Returns null when no caption is set.
+// The returned block has its trailing whitespace trimmed; the caller
+// must NOT pass it through `above` (which would re-pad it back out
+// to the table width). Append the lines directly instead.
+function _composeCaption(caption: string, width: number): Block | null {
+  if (caption === "") return null;
+  const centered = padLine(caption, width, "center").trimEnd();
+  return styled(Block.of(centered), { dim: true });
+}
+
+export function composeTable(node: LayoutNode): Block {
+  const { header, body, footer, columnCount } = _validateTable(node.attrs);
+  const attrs = node.attrs;
+  const cellPadding    = (attrs.cellPadding    as number)  ?? 1;
+  const columns        = (attrs.columns        as ColumnSpec[] | null) ?? [];
+  const columnDividers = (attrs.columnDividers as boolean) ?? true;
+  const headerDivider  = (attrs.headerDivider  as boolean) ?? true;
+  const footerDivider  = (attrs.footerDivider  as boolean) ?? true;
+  const rowDividers    = (attrs.rowDividers    as boolean) ?? false;
+  const borderStyleKey = resolveBorderStyle(
+    (attrs.borderStyle as string | undefined) ?? "rounded",
+  );
+  const ch          = BORDER_CHARS[borderStyleKey];
+  const borderColor = (attrs.borderColor as string | undefined) ?? "";
+  const titleColor  = (attrs.titleColor  as string | undefined) ?? "";
+  const title       = (attrs.title       as string | undefined) ?? "";
+  const caption     = (attrs.caption     as string | undefined) ?? "";
+  const wrapBorder  = styledWrapper(borderColor ? { fgColor: borderColor } : {});
+  const titleStyle: Style = titleColor ? { fgColor: titleColor } : {};
+
+  // Apply default-bold to header text cells before measurement so the
+  // bold SGR (zero visual width) doesn't perturb column sizing.
+  const styledHeader = header.map(_styleHeaderCell);
+
+  // Single measure pass over every row so columns line up across
+  // sections. We then slice rendered cells back into sections in the
+  // same order we assembled them.
+  const allRows: LayoutNode[][] = [
+    ...(styledHeader.length > 0 ? [styledHeader] : []),
+    ...body,
+    ...footer,
+  ];
+  const { layouts, cellBlocks } = _computeColumnLayouts(
+    allRows, columnCount, columns, cellPadding,
+  );
+  let idx = 0;
+  const headerCells = styledHeader.length > 0 ? [cellBlocks[idx++]] : [];
+  const bodyCells   = body.map(()   => cellBlocks[idx++]);
+  const footerCells = footer.map(() => cellBlocks[idx++]);
+
+  // A wide title may grow innerWidth beyond the cell grid; dividers and
+  // row wrappers both honour that final width so everything lines up.
+  const cellsWidth = _innerTableWidth(layouts, columnDividers);
+  const innerWidth = Math.max(
+    cellsWidth,
+    title !== "" ? minWidthForTitle(title) : 0,
+  );
+
+  // Build flat declarative list of section parts, then render each
+  // part into a Block in one place.
+  const sectionParts = _buildSectionParts(
+    headerCells, bodyCells, footerCells,
+    { headerDivider, footerDivider, rowDividers },
+  );
+  const renderPart = (part: SectionPart): Block =>
+    part.kind === "divider"
+      ? _composeDividerLine(layouts, columnDividers, innerWidth, ch, wrapBorder)
+      : _wrapRowSides(
+          _composeRowBlock(part.cells, layouts, columnDividers, ch.v, wrapBorder),
+          innerWidth, ch, wrapBorder,
+        );
+  const sectionBlocks = sectionParts.map(renderPart);
+
+  // Outer frame (top + bottom edges).
+  const topEdge    = Block.of(
+    buildTopEdge(ch, innerWidth, wrapBorder, title, titleStyle),
+  );
+  const bottomEdge = Block.of(
+    wrapBorder(ch.bl + ch.h.repeat(innerWidth) + ch.br),
+  );
+  const framed = [topEdge, ...sectionBlocks, bottomEdge]
+    .reduce(above, Block.empty());
+
+  // Optional caption below. Append directly via Block construction
+  // (not `above`) — `_composeCaption` returns a trimmed-trailing-space
+  // line, and routing it through `above` would re-pad it back out to
+  // the framed width.
+  const captionBlock = _composeCaption(caption, framed.width);
+  if (captionBlock === null) return framed;
+  return Block.of([...framed.lines, ...captionBlock.lines]);
+}
